+++ packages/polaris-core/src/config/create-apollo-config-util.ts
@@ -90,6 +90,7 @@ export function createPolarisMiddlewares(
     const middlewareConfiguration = config.middlewareConfiguration;
     if (config.supportedRealities) {
         const middlewaresMap = getMiddlewaresMap(
+            config,
             logger,
             config.supportedRealities,
             connectionManager,
@@ -130,6 +131,23 @@ export function createPolarisSubscriptionsConfig(config: PolarisServerConfig): a
     };
 }
 
+const mandatoryHeadersErrorMessage = (headers: any): string => {
+    const missingHeadersMessage = `Mandatory headers reality-id & requesting-sys are missing!`;
+    const missingRequestingSysMessage = `Mandatory header requesting-sys is missing!`;
+    const missingRealityIdMessage = `Mandatory header reality-id is missing!`;
+    const missingRealityId = headers[REALITY_ID] === undefined;
+    const missingRequestingSys = headers[REQUESTING_SYS] === undefined;
+    if (missingRealityId && missingRequestingSys) {
+        return missingHeadersMessage;
+    } else {
+        if (missingRealityId) {
+            return missingRealityIdMessage;
+        } else {
+            return missingRequestingSysMessage;
+        }
+    }
+};
+
 export function createPolarisContext(logger: AbstractPolarisLogger, config: PolarisServerConfig) {
     return (context: ExpressContext): PolarisGraphQLContext => {
         const { req, connection } = context;
@@ -137,10 +155,11 @@ export function createPolarisContext(logger: AbstractPolarisLogger, config: Pola
         const body = req ? req.body : connection;
 
         if (
+            body.operationName !== 'IntrospectionQuery' &&
             config.allowMandatoryHeaders &&
             (headers[REALITY_ID] === undefined || headers[REQUESTING_SYS] === undefined)
         ) {
-            const error = new Error('Mandatory headers were not set!');
+            const error = new Error(mandatoryHeadersErrorMessage(headers));
             logger.error(error.message);
             throw error;
         }