+++ packages/polaris-core/src/config/polaris-server-options.ts
@@ -27,5 +27,6 @@ export interface PolarisServerOptions extends Omit<ApolloServerExpressConfig, 'l
     connectionManager?: PolarisConnectionManager;
     enableFederation?: boolean;
     permissionsConfig?: PermissionsConfiguration;
+    enableDataVersionFilter?: boolean;
     connectionLessConfiguration?: ConnectionlessConfiguration;
 }