+++ packages/polaris-core/src/middlewares/middlewares-map.ts
@@ -8,8 +8,10 @@ import {
     SoftDeleteMiddleware,
 } from '@enigmatis/polaris-middlewares';
 import { PolarisConnectionManager } from '@enigmatis/polaris-typeorm';
+import { PolarisServerConfig } from '..';
 
 export const getMiddlewaresMap = (
+    config: PolarisServerConfig,
     logger: PolarisGraphQLLogger,
     realitiesHolder: RealitiesHolder,
     connectionManager?: PolarisConnectionManager,
@@ -18,6 +20,7 @@ export const getMiddlewaresMap = (
     const softDeleteMiddleware = new SoftDeleteMiddleware(logger).getMiddleware();
     const realitiesMiddleware = new RealitiesMiddleware(logger, realitiesHolder).getMiddleware();
     const dataVersionMiddleware = new DataVersionMiddleware(
+        config.enableDataVersionFilter,
         logger,
         realitiesHolder,
         connectionManager,