+++ packages/polaris-test/nest-server/graphql/entities/author.ts
@@ -1,6 +1,7 @@
 import { RepositoryEntity } from '@enigmatis/polaris-nest';
 import { Field, ObjectType } from '@nestjs/graphql';
 import { Book } from './book';
+import { Pen } from './pen';
 
 @ObjectType({
     implements: [RepositoryEntity],
@@ -14,4 +15,7 @@ export class Author extends RepositoryEntity {
 
     @Field(() => [Book], { nullable: true })
     public books: Book[];
+
+    @Field(() => [Pen], { nullable: true })
+    public pens: Pen[];
 }