+++ packages/polaris-test/server/schema/type-defs.ts
@@ -1,6 +1,7 @@
 export const typeDefs = `
     type Query {
         allBooks: [Book]!
+        authors: [Author]!
         allBooksPaginated: [Book]!
         allBooksWithWarnings: [Book]!
         authorById(id: String!): Author
@@ -12,10 +13,16 @@ export const typeDefs = `
         customContextInstanceMethod: String!
         permissionsField: String @permissions(entityTypes: ["foo"], actions: ["READ", "DELETE"])
     }
-
+    input ReviewKind{
+        site: String
+        name: String
+    }
     type Mutation {
         createAuthor(firstName: String!, lastName: String): Author!
         createBook(title: String!, authorId: String): Book!
+        createPen(color: String!, id: String): Pen!
+        createChapter(number: Int!, bookId: String): Chapter!
+        createReview(description:String!, rating:String!, bookId: String!, reviewKind: ReviewKind!): Review!
         updateBooksByTitle(title: String!, newTitle: String!): [Book]!
         deleteBook(id: String!): Boolean
         deleteAuthor(id: String!): Boolean
@@ -36,8 +43,70 @@ export const typeDefs = `
         realityId: Int!
         title: String
         author: Author
+        chapters: [Chapter]
+        reviews: [Review]
+    }
+    interface Review implements RepositoryEntity {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        rating: Int!
+        description: String!
+        book: Book!
+    }
+    type ProfessionalReview implements Review {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        rating: Int!
+        description: String!
+        book: Book!
+        site: String!
+    }
+    type SimpleReview implements Review {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        rating: Int!
+        description: String!
+        book: Book!
+        name: String!
     }
 
+    type Pen implements RepositoryEntity {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        color: String
+        author: Author
+    }
+    type Chapter implements RepositoryEntity {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        number: Int!
+        book: Book
+    }
     type Author implements RepositoryEntity {
         id: String!
         deleted: Boolean!
@@ -49,5 +118,6 @@ export const typeDefs = `
         firstName: String
         lastName: String
         books: [Book]
+        pens: [Pen]
     }
 `;