+++ packages/polaris-typeorm/src/handlers/data-version-handler.ts
@@ -1,5 +1,8 @@
 import { PolarisExtensions, PolarisGraphQLContext } from '@enigmatis/polaris-common';
-import { DataVersion, PolarisConnection, QueryRunner } from '..';
+import { RelationMetadata } from 'typeorm/metadata/RelationMetadata';
+import { DataVersion, PolarisConnection, QueryRunner, SelectQueryBuilder } from '..';
+import { EntityMetadata } from 'typeorm';
+import { isDescendentOfCommonModel } from '../utils/descendent-of-common-model';
 
 export class DataVersionHandler {
     public async updateDataVersion<Entity>(
@@ -66,3 +69,96 @@ export class DataVersionHandler {
         return result;
     }
 }
+
+const extractRelations = (
+    dvMapping: Map<any, any>,
+    entityMetadata: EntityMetadata,
+    relation: RelationMetadata,
+) => {
+    let children;
+    if (dvMapping.size > 1) {
+        if (dvMapping.has(entityMetadata.name)) {
+            children = dvMapping.get(entityMetadata.name);
+        } else {
+            if (dvMapping.has(relation.inverseSidePropertyPath)) {
+                children = dvMapping.get(relation.inverseSidePropertyPath);
+            }
+        }
+    } else {
+        children = dvMapping.entries().next().value[1];
+    }
+    return children;
+};
+
+function getPropertyMap(propertyRelations: any | any[], relation: RelationMetadata) {
+    return propertyRelations instanceof Array
+        ? propertyRelations.find((x: any) => x.has(relation.propertyName))
+        : propertyRelations.has(relation.propertyName)
+        ? propertyRelations
+        : undefined;
+}
+
+function leftJoinRelationsToQB(
+    children: any,
+    relation: RelationMetadata,
+    names: string[],
+    qb: any,
+    entityMetadata: EntityMetadata,
+) {
+    const childDVMapping = getPropertyMap(children, relation);
+    if (childDVMapping) {
+        const relationMetadata = relation.inverseEntityMetadata;
+        const alias: string = relationMetadata.tableName;
+        const notInJoins = names.filter((x) => x === alias).length === 0;
+        if (isDescendentOfCommonModel(relationMetadata) && notInJoins) {
+            qb = qb.leftJoinAndSelect(
+                `${entityMetadata.tableName}.${relation.propertyName}`,
+                alias,
+            );
+            names.push(alias);
+            qb = loadRelations(qb, relationMetadata, names, childDVMapping);
+        }
+    }
+    return qb;
+}
+
+const loadRelations = (
+    qb: any,
+    entityMetadata: EntityMetadata,
+    names: string[],
+    dvMapping: Map<any, any>,
+): any => {
+    if (entityMetadata.relations && dvMapping.size > 0) {
+        for (const relation of entityMetadata.relations) {
+            const children = extractRelations(dvMapping, entityMetadata, relation);
+            if (children) {
+                qb = leftJoinRelationsToQB(children, relation, names, qb, entityMetadata);
+            }
+        }
+    }
+    return qb;
+};
+
+export const dataVersionFilter = (
+    connection: PolarisConnection,
+    qb: SelectQueryBuilder<any>,
+    entityName: string,
+    context: PolarisGraphQLContext,
+    shouldLoadRelations: boolean,
+) => {
+    if (context.requestHeaders.dataVersion && context.requestHeaders.dataVersion > 0) {
+        qb = qb.distinct();
+        const entityMetadata = connection.getMetadata(entityName);
+        let names = [entityName];
+        if (context.dataVersionContext?.mapping && shouldLoadRelations) {
+            qb = loadRelations(qb, entityMetadata, names, context.dataVersionContext!.mapping!);
+        }
+        const dataVersion = context.requestHeaders.dataVersion;
+        qb.where(`${qb.alias}.dataVersion > :dataVersion`, { dataVersion });
+        names = names.slice(1);
+        for (const name of names) {
+            qb = qb.orWhere(`${name}.dataVersion > :dataVersion`);
+        }
+    }
+    return qb;
+};