+++ packages/polaris-typeorm/src/handlers/find-handler.ts
@@ -1,5 +1,5 @@
 import { PolarisRequestHeaders } from '@enigmatis/polaris-common';
-import { FindConditions, In, MoreThan } from 'typeorm';
+import { FindConditions, In } from 'typeorm';
 import { PolarisFindManyOptions, PolarisFindOneOptions } from '..';
 
 export const getEntitiesIncludingDeletedConditions: FindConditions<any> = {
@@ -15,6 +15,7 @@ export class FindHandler {
     public findConditions<Entity>(
         includeLinkedOper: boolean,
         polarisOptions?: PolarisFindManyOptions<Entity> | PolarisFindOneOptions<Entity>,
+        shouldIncludeDeletedEntities: boolean = false,
     ) {
         const headers: PolarisRequestHeaders = polarisOptions?.context?.requestHeaders || {};
 
@@ -28,11 +29,8 @@ export class FindHandler {
         }
 
         polarisCriteria.where = { ...polarisCriteria.where };
-        if (polarisCriteria.where.deleted === undefined) {
+        if (polarisCriteria.where.deleted === undefined && !shouldIncludeDeletedEntities) {
             polarisCriteria.where.deleted = false;
-        }
-        if (polarisCriteria.where.dataVersion === undefined && headers.dataVersion) {
-            polarisCriteria.where.dataVersion = MoreThan(headers.dataVersion);
         }
         if (polarisCriteria.where.realityId === undefined) {
             polarisCriteria.where.realityId = realityIdCriteria(includeLinkedOper, headers);