+++ packages/polaris-typeorm/src/typeorm-bypasses/polaris-repository.ts
@@ -7,29 +7,26 @@ import {
     FindOneOptions,
     ObjectID,
     ObjectLiteral,
+    QueryRunner,
     Repository,
     SaveOptions,
+    SelectQueryBuilder,
     UpdateResult,
 } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
 import {
-    CommonModel,
     PolarisCriteria,
+    PolarisEntityManager,
     PolarisFindManyOptions,
     PolarisFindOneOptions,
     PolarisSaveOptions,
 } from '..';
+import { isDescendentOfCommonModel } from '../utils/descendent-of-common-model';
 
 /**
  * Repository is supposed to work with your entity objects. Find entities, insert, update, delete, etc.
  */
 export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<Entity> {
-    public isDescendentOfCommonModel(): boolean {
-        return (
-            this.metadata.inheritanceTree[this.metadata.inheritanceTree.length - 1].name ===
-            CommonModel.name
-        );
-    }
     /**
      * Saves one or many given entities.
      */
@@ -42,7 +39,7 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<T | T[]> {
         return this.manager.save<T>(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisSaveOptions(entityOrEntities, context) as any)
                 : entityOrEntities,
             options,
@@ -73,7 +70,7 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<UpdateResult> {
         return this.manager.update(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisCriteria(criteria, context) as any)
                 : criteria,
             partialEntity,
@@ -102,7 +99,7 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<DeleteResult> {
         return this.manager.delete(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisCriteria(criteria, context) as any)
                 : criteria,
         );
@@ -118,7 +115,7 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<number> {
         return this.manager.count(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisFindManyOptions(optionsOrConditions, context) as any)
                 : optionsOrConditions,
         );
@@ -134,7 +131,7 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<Entity[]> {
         return this.manager.find(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisFindManyOptions(optionsOrConditions, context) as any)
                 : optionsOrConditions,
         );
@@ -157,10 +154,44 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
     ): Promise<Entity | undefined> {
         return this.manager.findOne(
             this.metadata.target as any,
-            this.isDescendentOfCommonModel()
+            isDescendentOfCommonModel(this.metadata)
                 ? (new PolarisFindOneOptions(optionsOrConditions, context) as any)
                 : optionsOrConditions,
             maybeOptions,
+        );
+    }
+
+    /**
+     * Creates a new query builder that can be used to build a sql query.
+     */
+    // @ts-ignore
+    public createQueryBuilder(
+        context: PolarisGraphQLContext,
+        alias?: string,
+        queryRunner?: QueryRunner,
+    ): SelectQueryBuilder<Entity> {
+        return ((this.manager as unknown) as PolarisEntityManager).createQueryBuilder<Entity>(
+            this.metadata.target as any,
+            alias,
+            queryRunner || this.queryRunner,
+            undefined,
+            context,
+            false,
+        );
+    }
+
+    public createQueryBuilderWithDeletedEntities(
+        context: PolarisGraphQLContext,
+        alias?: string,
+        queryRunner?: QueryRunner,
+    ): SelectQueryBuilder<Entity> {
+        return ((this.manager as unknown) as PolarisEntityManager).createQueryBuilder<Entity>(
+            this.metadata.target as any,
+            alias,
+            queryRunner || this.queryRunner,
+            undefined,
+            context,
+            true,
         );
     }
 }