+++ packages/polaris-test/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-test@1.0.3...@enigmatis/polaris-test@1.1.0) (2020-10-21)
+
+
+### Bug Fixes
+
+* informative error on mandatory headers, and make it work on playground ([#29](https://github.com/Enigmatis/polaris-united/issues/29)) ([39ec502](https://github.com/Enigmatis/polaris-united/commit/39ec50256a74f6a53bd8e52c88b1993ff30e406c))
+
+
+### Features
+
+* added support for query builder ([#28](https://github.com/Enigmatis/polaris-united/issues/28)) ([31da6dd](https://github.com/Enigmatis/polaris-united/commit/31da6dd71b5a21995fc37414705ddcb89dbdc6aa))
+* dv by child entity ([7556e71](https://github.com/Enigmatis/polaris-united/commit/7556e7145375bd91cac0efe1f7d4592e0f0d9091))
+
+
+
+
+
 ## [1.0.3](https://github.com/Enigmatis/polaris-united/compare/@enigmatis/polaris-test@1.0.2...@enigmatis/polaris-test@1.0.3) (2020-10-12)
 
 **Note:** Version bump only for package @enigmatis/polaris-test